streamline code and restore missing files
